refactor: Convert homepage to Server Component
